move type declaration
